+++ packages/runtime/src/lib/txn.ts
@@ -220,6 +220,13 @@ export function txAppArg(
 ): StackElem {
 	const tealVersion: number = interpreter.tealVersion;
 
+	if (txField === "Logs") {
+		const txReceipt = interpreter.runtime.ctx.state.txReceipts.get(tx.txID);
+		const logs: Uint8Array[] = txReceipt?.logs ?? [];
+		op.checkIndexBound(idx, logs, op.line);
+		return parseToStackElem(logs[idx], txField);
+	}
+
 	const s = TxnFields[tealVersion][txField]; // 'apaa' or 'apat'
 	const result = tx[s as keyof EncTx] as Buffer[]; // array of pk buffers (accounts or appArgs)
 	if (!result) {