+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -6897,6 +6897,42 @@ describe("Teal Opcodes", function () {
 		});
 	});
 
+	describe("Logs", function () {
+		let stack: Stack<StackElem>;
+		let interpreter: Interpreter;
+		this.beforeEach(() => {
+			stack = new Stack<StackElem>();
+			interpreter = new Interpreter();
+			interpreter.runtime = new Runtime([]);
+			interpreter.runtime.ctx.tx = TXN_OBJ;
+			interpreter.tealVersion = 6;
+			interpreter.innerTxnGroups = [[TXN_OBJ, { ...TXN_OBJ, fee: 1000 }]];
+			interpreter.runtime.ctx.state.txReceipts.set(TXN_OBJ.txID, {
+				txn: interpreter.runtime.ctx.tx,
+				txID: TXN_OBJ.txID,
+				logs: [parsing.stringToBytes("Monty"), parsing.stringToBytes("Python")],
+			});
+		});
+
+		it("Should put on top of the stack log from group transaction", () => {
+			const op = new Gitxna(["1", "Logs", "0"], 1, interpreter);
+			op.execute(stack);
+			assert.deepEqual(stack.pop(), parsing.stringToBytes("Monty"));
+		});
+
+		it("Should throw an error index out of bound", () => {
+			const op = new Gitxna(["1", "Logs", "2"], 1, interpreter);
+			expectRuntimeError(() => op.execute(stack), RUNTIME_ERRORS.TEAL.INDEX_OUT_OF_BOUND);
+		});
+
+		it("Should put on top of stack log from group transaction", () => {
+			stack.push(1n);
+			const op = new Gitxnas(["1", "Logs"], 1, interpreter);
+			op.execute(stack);
+			assert.deepEqual(stack.pop(), parsing.stringToBytes("Python"));
+		});
+	});
+
 	describe("Tealv7: base64Decode opcode", function () {
 		let stack: Stack<StackElem>;
 		const encoded64BaseStd = "YWJjMTIzIT8kKiYoKSctPUB+";